Moved ConnectionHandshakeManager, NetPeerManager into their own files. Misc doxygen comments too
