+++ Documentation/ReadMe.txt
@@ -1,7 +1,2 @@
-In order to read the documentation you may have to unblock it by doing this:
-
-1. Right-click the .CHM file and choose Properties
-2. On the General tab, click the button labeled "Unblock"
-3. Click OK
-
-For reference, see: http://support.microsoft.com/kb/2021383
+sudo apt install libssl-dev zlib1g-dev libncurses5-dev libncursesw5-dev libreadline-dev libsqlite3-dev 
+sudo apt install libgdbm-dev libdb5.3-dev libbz2-dev libexpat1-dev liblzma-dev tk-dev
