+++ Lidgren.Network/Lidgren.Network.csproj
@@ -113,6 +113,7 @@
     <Compile Include="NetPeer.MessagePools.cs" />
     <Compile Include="NetPeer.Send.cs" />
     <Compile Include="NetPeerConfiguration.cs" />
+    <Compile Include="NetPeerManager.cs" />
     <Compile Include="NetPeerStatistics.cs" />
     <Compile Include="NetPeerStatus.cs" />
     <Compile Include="NetQueue.cs" />
@@ -164,6 +165,9 @@
       <ProductName>Windows Installer 3.1</ProductName>
       <Install>true</Install>
     </BootstrapperPackage>
+  </ItemGroup>
+  <ItemGroup>
+    <Compile Include="ConnetionHandshakeManager.cs" />
   </ItemGroup>
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
   <!-- To modify your build process, add your task inside one of the targets below and uncomment it.
