+++ Lidgren.Network/NetPeer.Internal.cs
@@ -15,234 +15,6 @@ using NetEndPoint = System.Net.IPEndPoint;
 
 namespace Lidgren.Network
 {
-    public static class NetPeerManager
-    {
-        /// <summary>
-        /// Peer List (thread safe)
-        /// </summary>
-        public static ConcurrentDictionary<NetPeer, Socket> Peers = new ConcurrentDictionary<NetPeer, Socket>();
-
-        public static Thread NetworkUpdateThread = null;
-        public static Object InitThreadLock = new Object();
-
-        /// <summary>
-        /// Wait for network thread exit
-        /// </summary>
-        public static void WaitForExit()
-        {
-            NetworkUpdateThread?.Join();
-        }
-        
-        /// <summary>
-        /// Concurrent add peer task
-        /// </summary>
-        /// <param name="peer"></param>
-        public static void AddPeer(NetPeer peer)
-        {
-            if (!Peers.TryAdd(peer, peer.Socket))
-            {
-                Task.Run(() => AddPeerTask(peer));
-            }
-        }
-        
-        /// <summary>
-        /// Concurrent remove peer task
-        /// </summary>
-        /// <param name="peer"></param>
-        public static void RemovePeer(NetPeer peer)
-        {
-            if (Peers.TryRemove(peer, out _))
-            {
-                Task.Run(() => RemovePeerTask(peer));
-            }
-        }
-
-        // pattern is, if you can't do it now, do it later.
-
-
-        /// <summary>
-        /// Add peer task, will add when succedes
-        /// </summary>
-        /// <param name="peer"></param>
-        private static async Task AddPeerTask(NetPeer peer)
-        {
-            // return immediately
-            if (Peers == null) return;
-
-            // wait for add to work
-            while (!Peers.TryAdd(peer, peer.Socket))
-            {
-                await Task.Delay(1);
-
-                // wait until add succedes
-            }
-        }
-
-
-        /// <summary>
-        /// Remove task, will add when succedes
-        /// </summary>
-        /// <param name="peer"></param>
-        private static async Task RemovePeerTask(NetPeer peer)
-        {
-            // return immediately
-            if (Peers == null) return;
-
-            // wait for remove to work
-            while (!Peers.TryRemove(peer, out _))
-            {
-                await Task.Delay(1);
-
-                // wait until add succedes
-            }
-        }
-
-
-        public static void StartNetworkThread()
-        {
-            lock (InitThreadLock)
-            {
-                // thread shutdown behaviour is, it will shut down when netpeer's have been shut down, if it suddenly gets a new netpeer to work with
-                // it initialises a new thread for it to use
-                // if the thread has died, or the thread hasn't been started, overwrite it.
-                if (NetworkUpdateThread == null || NetworkUpdateThread.IsAlive == false)
-                {
-                    // extra alloc, just in case it's a dead thread
-                    NetworkUpdateThread = null;
-
-                    // start network thread
-                    NetworkUpdateThread = new Thread(NetworkLoop)
-                    {
-                        Name = "Lidgren.Network Thread"
-                    };
-
-                    //m_networkThread.IsBackground = true;
-                    NetworkUpdateThread.Start();
-                }
-            }
-        }
-
-        /// <summary>
-        /// Singular network update thread
-        /// This must be static, because otherwise people might try to use peer in a local thread context which is incorrect.
-        /// </summary>
-        private static void NetworkLoop()
-        {
-            bool running = true;
-            bool hadElements = false;
-            while (running)
-            {
-                // if our peer manager had elements and they're now gone, time to shutdown the thread, if we have no peers left
-                if (hadElements && running)
-                {
-                    if (Peers.Count == 0)
-                    {
-                        return; // time to exit, no peers exist anymore.
-                    }
-                }
-
-                foreach (var kvp in Peers)
-                {
-                    hadElements = true;
-                    var peer = kvp.Key;
-                    if (peer.Status == NetPeerStatus.Running)
-                    {
-                        peer.PeerUpdate();
-                    }
-
-                    if (peer.Status == NetPeerStatus.ShutdownRequested)
-                    {
-                        try
-                        {
-                            // shutdown peer
-                            peer.ExecutePeerShutdown();
-                        }
-                        catch (Exception e)
-                        {
-                            Console.WriteLine(e);
-                            throw;
-                        }
-                    }
-                }
-
-                // no sockets to poll prep for exit condition
-                if (Peers.Count == 0) continue;
-
-                // Socket.Select is faster than running Socket.Poll foreach socket.
-                var sockets = Peers.Values.ToList();
-                Socket.Select(sockets, null, null, 1000);
-
-                // only update selected sockets
-                foreach (var socket in sockets)
-                {
-                    // retrieve socket peer to process data on it, as data is pending on it's connection
-                    var pair = Peers.SingleOrDefault(kvp => kvp.Value == socket);
-
-                    // potential place to pass to socket Task?
-                    pair.Key.SocketDataHandler();
-                }
-
-            }
-        }
-    }
-
-    public class ConnectionHandshakeManager
-    {
-        public ConcurrentDictionary<NetEndPoint, NetConnection> Handshakes = new ConcurrentDictionary<IPEndPoint, NetConnection>();
-
-        public void AddHandshake(NetEndPoint endPoint, NetConnection connection)
-        {
-            if (!Handshakes.TryAdd(endPoint, connection))
-            {
-                Task.Run(() => AddHandshakeTask(endPoint, connection));
-            }
-        }
-
-        public void RemoveHandshake(NetEndPoint endPoint)
-        {
-            if (!Handshakes.TryRemove(endPoint, out _))
-            {
-                Task.Run(() => RemoveHandshakeTask(endPoint));
-            }
-        }
-
-        /// <summary>
-        /// Add peer task, will add when succedes
-        /// </summary>
-        /// <param name="peer"></param>
-        private async Task AddHandshakeTask(NetEndPoint endPoint, NetConnection connection)
-        {
-            // return immediately
-            if (Handshakes == null) return;
-
-            // wait for add to work
-            while (!Handshakes.TryAdd(endPoint, connection))
-            {
-                await Task.Delay(1);
-
-                // wait until add succedes
-            }
-        }
-
-        /// <summary>
-        /// Add peer task, will add when succedes
-        /// </summary>
-        /// <param name="peer"></param>
-        private async Task RemoveHandshakeTask(NetEndPoint endPoint)
-        {
-            // return immediately
-            if (Handshakes == null) return;
-
-            // wait for add to work
-            while (!Handshakes.TryRemove(endPoint, out _))
-            {
-                await Task.Delay(1);
-                if (!Handshakes.ContainsKey(endPoint)) return;
-                // wait until add succedes
-            }
-        }
-    }
-
     public partial class NetPeer
     {
         private NetPeerStatus m_status = NetPeerStatus.NotRunning;
@@ -618,7 +390,9 @@ namespace Lidgren.Network
             }
         }
 
-
+        /// <summary>
+        /// Peer update
+        /// </summary>
         public void PeerUpdate()
         {
             int maxCHBpS = 1250;
