+++ Lidgren.Network/NetPeer.cs
@@ -126,8 +126,6 @@ namespace Lidgren.Network
 		}
 
         
-
-
 		/// <summary>
 		/// Binds to socket and spawns the networking thread
 		/// </summary>
@@ -211,16 +209,16 @@ namespace Lidgren.Network
 			return retval;
 		}
 		
-        	/// <summary>
-	        /// Reads a pending message from any connection, if any.
-	        /// Returns true if message was read, otherwise false.
-	        /// </summary>
-	        /// <returns>True, if message was read.</returns>
-	        public bool ReadMessage(out NetIncomingMessage message)
-	        {
-	            message = ReadMessage();
-	            return message != null;
-	        }
+		/// <summary>
+		/// Reads a pending message from any connection, if any.
+		/// Returns true if message was read, otherwise false.
+		/// </summary>
+		/// <returns>True, if message was read.</returns>
+		public bool ReadMessage(out NetIncomingMessage message)
+		{
+			message = ReadMessage();
+			return message != null;
+		}
 
 		/// <summary>
 		/// Read a pending message from any connection, if any
